Added CamsRole enum and mapped to mock users
Jira ticket: CAMS-406
